content updated

+++ Main.py
@@ -49,7 +49,7 @@ screen_height =window.winfo_screenheight()
 
 
 x = int((screen_width/2) -(window_width/2))
-y =(screen_height/2) - (window_height/2)
+y =int((screen_height/2) - (window_height/2))
 
 window.geometry(f"{window_width}x{window_height}+{x}+{y}")
